Move k8s code inside COCommon Orchestrator and split topology interface
