+++ pkg/csi/service/vanilla/controller.go
@@ -44,10 +44,10 @@ import (
 	"sigs.k8s.io/vsphere-csi-driver/pkg/common/utils"
 	"sigs.k8s.io/vsphere-csi-driver/pkg/csi/service/common"
 	"sigs.k8s.io/vsphere-csi-driver/pkg/csi/service/common/commonco"
+	commoncotypes "sigs.k8s.io/vsphere-csi-driver/pkg/csi/service/common/commonco/types"
 	"sigs.k8s.io/vsphere-csi-driver/pkg/csi/service/logger"
 	csitypes "sigs.k8s.io/vsphere-csi-driver/pkg/csi/types"
 	"sigs.k8s.io/vsphere-csi-driver/pkg/internalapis/cnsvolumeoperationrequest"
-	csinodetopology "sigs.k8s.io/vsphere-csi-driver/pkg/internalapis/csinodetopology"
 )
 
 // NodeManagerInterface provides functionality to manage (VM) nodes.
@@ -65,7 +65,7 @@ type controller struct {
 	manager     *common.Manager
 	nodeMgr     NodeManagerInterface
 	authMgr     common.AuthorizationService
-	topologyMgr csinodetopology.TopologyService
+	topologyMgr commoncotypes.ControllerTopologyService
 }
 
 // volumeMigrationService holds the pointer to VolumeMigration instance.
@@ -234,7 +234,7 @@ func (c *controller) Init(config *cnsconfig.Config, version string) error {
 	// Create dynamic informer for CSINodeTopology instance if FSS is enabled.
 	if commonco.ContainerOrchestratorUtility.IsFSSEnabled(ctx, common.ImprovedVolumeTopology) {
 		// Initialize volume topology service.
-		c.topologyMgr, err = csinodetopology.InitTopologyServiceInterface(ctx)
+		c.topologyMgr, err = commonco.ContainerOrchestratorUtility.InitTopologyServiceInController(ctx)
 		if err != nil {
 			log.Errorf("failed to initialize topology service. Error: %+v", err)
 			return err
@@ -426,21 +426,26 @@ func (c *controller) createBlockVolume(ctx context.Context, req *csi.CreateVolum
 		if commonco.ContainerOrchestratorUtility.IsFSSEnabled(ctx, common.ImprovedVolumeTopology) {
 			sharedDatastores, datastoreTopologyMap, err = c.topologyMgr.GetSharedDatastoresInTopology(ctx,
 				topologyRequirement)
+			if err != nil || len(sharedDatastores) == 0 {
+				return nil, logger.LogNewErrorCodef(log, codes.Internal,
+					"failed to get shared datastores for topology requirement: %+v. Error: %+v",
+					topologyRequirement, err)
+			}
 		} else {
 			if c.manager.CnsConfig.Labels.Zone == "" || c.manager.CnsConfig.Labels.Region == "" {
 				// If zone and region label (vSphere category names) not specified in
 				// the config secret, then return NotFound error.
-				return nil, logger.LogNewErrorCode(log, codes.NotFound,
+				return nil, logger.LogNewErrorCode(log, codes.Internal,
 					"zone/region vsphere category names not specified in the vsphere config secret")
 			}
 			vcenter, err := c.manager.VcenterManager.GetVirtualCenter(ctx, c.manager.VcenterConfig.Host)
 			if err != nil {
-				return nil, logger.LogNewErrorCodef(log, codes.NotFound,
+				return nil, logger.LogNewErrorCodef(log, codes.Internal,
 					"failed to get vCenter. Err: %v", err)
 			}
 			tagManager, err := cnsvsphere.GetTagManager(ctx, vcenter)
 			if err != nil {
-				return nil, logger.LogNewErrorCodef(log, codes.NotFound,
+				return nil, logger.LogNewErrorCodef(log, codes.Internal,
 					"failed to get tagManager. Err: %v", err)
 			}
 			defer func() {
@@ -452,11 +457,11 @@ func (c *controller) createBlockVolume(ctx context.Context, req *csi.CreateVolum
 			sharedDatastores, datastoreTopologyMap, err = c.nodeMgr.GetSharedDatastoresInTopology(ctx,
 				topologyRequirement, tagManager, c.manager.CnsConfig.Labels.Zone, c.manager.CnsConfig.Labels.Region)
 			if err != nil || len(sharedDatastores) == 0 {
-				return nil, logger.LogNewErrorCodef(log, codes.NotFound,
+				return nil, logger.LogNewErrorCodef(log, codes.Internal,
 					"failed to get shared datastores in topology: %+v. Error: %+v", topologyRequirement, err)
 			}
 		}
-		log.Debugf("Shared datastores [%+v] retrieved for topologyRequirement [%+v] with " +
+		log.Debugf("Shared datastores [%+v] retrieved for topologyRequirement [%+v] with "+
 			"datastoreTopologyMap [+%v]", sharedDatastores, topologyRequirement, datastoreTopologyMap)
 		if createVolumeSpec.ScParams.DatastoreURL != "" {
 			// Check datastoreURL specified in the storageclass is accessible from
